⚙️ HUGE CHANGES

+++ src/events/interactionCreate.ts
@@ -1,13 +1,16 @@
 import type { Event, InteractionCommand } from '../@types';
+import * as Util from '../utils/functions';
 import * as Items from '../database/rpg/Items';
 import * as Emojis from '../emojis.json';
-
 import InteractionCommandContext from '../structures/Interaction';
 
 export const name: Event["name"] = "interactionCreate";
 export const execute: Event["execute"] = async (interaction: InteractionCommand) => {
     if (!interaction.isCommand()) return;
     if (!interaction.client._ready) return interaction.reply({ content: "The bot is still loading, please wait a few seconds and try again."});
+    if (await interaction.client.database.getCooldownCache(interaction.user.id)) return interaction.reply({
+        content: (interaction.client.translations.get('en-US')("base:COOLDOWN")).replace("{{emojis.jolyne}}", Emojis.jolyne)
+    });
 
     const command = interaction.client.commands.get(interaction.commandName);
     if (!command) return;
@@ -48,21 +51,47 @@ export const execute: Event["execute"] = async (interaction: InteractionCommand)
 
         }
         if (hasChanged) interaction.client.database.saveUserData(userData);
-    
-        await command.execute(new InteractionCommandContext(interaction), userData);
 
+        if (command.rpgCooldown) {
+            const cd = parseInt(await interaction.client.database.redis.client.get(`jjba:rpg_cooldown_${interaction.user.id}:${command.name}`));
+            if (cd && cd > Date.now()) return interaction.reply({ content: interaction.client.translations.get("en-US")(command.rpgCooldown.i18n ?? 'base:RPG_COOLDOWN', {
+                time: Util.generateDiscordTimestamp(cd, 'REMAINS')
+            })});
+            await interaction.client.database.redis.client.set(`jjba:rpg_cooldown_${interaction.user.id}:${command.name}`, Date.now() + command.rpgCooldown["base"]);
+        }
+            
+    
+        const ctx = new InteractionCommandContext(interaction)
+        await command.execute(ctx, userData);
+
+        await Util.wait(2000);
         // Misc
         if (interaction.replied) {
             const newMails = await interaction.client.database.redis.client.get(`jjba:newUnreadMails:${interaction.user.id}`);
             if (newMails) {
                 interaction.client.database.redis.client.del(`jjba:newUnreadMails:${interaction.user.id}`);
                 interaction.followUp({
-                    content: interaction.client.translations.get(userData.language)("base:NEW_MAILS", {
+                    content: ctx.translate("base:NEW_MAILS", {
                         count: newMails
                     })
                 });
-
             }
+            if (await interaction.client.database.getCooldownCache(interaction.user.id)) return;
+            checkLVL()
+            function checkLVL() {
+                if (userData.xp >= Util.getMaxXp(userData.level)) {
+                    userData.xp = userData.xp - Util.getMaxXp(userData.level);
+                    userData.level++;
+                    ctx.followUp({
+                        content: ctx.translate("base:LEVEL_UP_MESSAGE", {
+                            level: userData.level
+                        })
+                    });
+                    ctx.client.database.saveUserData(userData);
+                    if (userData.xp >= Util.getMaxXp(userData.level)) checkLVL();
+                }
+            }
+
 
         }
     } else command.execute(new InteractionCommandContext(interaction));
